Move AppIcon enum to its own file, update comments and clean up code
