+++ scripts/update-icons.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env sh
 
-### Icon Composer Tool
 #
 # Usage:
 #
@@ -15,11 +14,7 @@
 #
 #     ictool input-document.icon --export-preview iOS Light 1024 1024 2 output.png
 #
-#
-###
 alias ictool="/Applications/Xcode26-beta7.app/Contents/Applications/Icon\ Composer.app/Contents/Executables/ictool"
-
-# ictool input-document.icon --export-preview iOS Light 1024 1024 2 output.png
 
 # create an array of all source Icon Composer files
 icon_composer_files=(BikeIndex/AppIcons/*.icon)
@@ -29,9 +24,5 @@ for ((i=0; i<${#icon_composer_files[@]}; i++)); do
     icon_name=`echo ${icon_composer_files[$i]} | cut -d '/' -f 3 | sed 's/.icon//'`
     output_path="BikeIndex/Assets.xcassets/AppIcons-in-app/${icon_name}-in-app.imageset/AppIcon.jpg"
 
-    # echo "${icon_composer_files[$i]}" "$output_path"
     ictool "${icon_composer_files[$i]}" --export-preview iOS Light 1024 1024 2 "$output_path"
 done
-
-
-# Assets.xcassets/AppIcons-in-app/Blue-on-inverse-in-app.imageset/AppIcon.jpg
